Add test/expect.js and separate the generated tests from the reference image ones
